proposal.go proposal_data.go serializable_proposal.go save_load_test.go
